requirements.txt
Django==1.8
django-filter==0.12.0
djangorestframework==3.3.2
Markdown==2.6.5
Pillow==3.1.1
wheel==0.26.0
gunicorn
whitenoise
dj-database-url==0.2.1